Optimization - use float instead of double

+++ src/SpiralWatchy.h
@@ -21,17 +21,17 @@ public:
                               int16_t w, int16_t h, uint16_t color, bool even);
 
 
-  void DrawHand(double angle, double size);                           
+  void DrawHand(float angle, float size);                           
   void fillTriangle(VectorInt v0, Vector uv0, VectorInt v1, Vector uv1, VectorInt v2, Vector uv2, const uint8_t bitmap[], int w, int h);
   
-   void drawLine(int x, int y, int w, VectorInt v0, Vector uv0, VectorInt a, Vector uv1, VectorInt b, Vector uv2, double den, const uint8_t *bitmap, int16_t bw, int16_t bh);
+   void drawLine(int x, int y, int w, VectorInt v0, Vector uv0, VectorInt a, Vector uv1, VectorInt b, Vector uv2, float den, const uint8_t *bitmap, int16_t bw, int16_t bh);
 
                                 
   void writeFastHLineUV(int16_t x, int16_t y, int16_t w, Vector uvA, Vector uvB, const uint8_t bitmap[], int bw, int bh);
 
   void fillTriangle2(VectorInt v0, Vector uv0, VectorInt v1, Vector uv1, VectorInt v2, Vector uv2, const uint8_t bitmap[], int w, int h);
   
-  void drawLine2(int x, int y, int w, VectorInt v0, Vector uv0, VectorInt a, Vector uv1, VectorInt b, Vector uv2, double den, const uint8_t *bitmap, int16_t bw, int16_t bh);
+  void drawLine2(int x, int y, int w, VectorInt v0, Vector uv0, VectorInt a, Vector uv1, VectorInt b, Vector uv2, float den, const uint8_t *bitmap, int16_t bw, int16_t bh);
             
   void writeFastHLineUV2(int16_t x, int16_t y, int16_t w, Vector uvA, Vector uvB, const uint8_t bitmap[], int bw, int bh);
 
@@ -39,7 +39,7 @@ public:
 
   void fillTriangle(VectorInt v0, Vector uv0, VectorInt v1, Vector uv1, VectorInt v2, Vector uv2, const uint8_t bitmap[], int w, int h, uint16_t color);
   
-  void drawLine(int x, int y, int w, VectorInt v0, Vector uv0, VectorInt a, Vector uv1, VectorInt b, Vector uv2, double den, const uint8_t *bitmap, int16_t bw, int16_t bh, uint16_t color);
+  void drawLine(int x, int y, int w, VectorInt v0, Vector uv0, VectorInt a, Vector uv1, VectorInt b, Vector uv2, float den, const uint8_t *bitmap, int16_t bw, int16_t bh, uint16_t color);
             
   void writeFastHLineUV(int16_t x, int16_t y, int16_t w, Vector uvA, Vector uvB, const uint8_t bitmap[], int bw, int bh, uint16_t color);
 };
