+++ src/Vector.h
@@ -5,16 +5,16 @@
 
 struct Vector
 {
-    double x;
-    double y;
+    float x;
+    float y;
 
 public:
-    static double dotProduct(const Vector& v1, const Vector& v2)
+    static float dotProduct(const Vector& v1, const Vector& v2)
     {
         return v1.x * v2.x + v1.y * v2.y;
     }
 
-    static double crossProduct(const Vector& v1, const Vector& v2)
+    static float crossProduct(const Vector& v1, const Vector& v2)
     {
         return v1.x * v2.y - v1.y * v2.x;
     }
@@ -36,11 +36,11 @@ public:
         }
     }
 
-    static Vector rotateVector(const Vector& vector, const double& angle)
+    static Vector rotateVector(const Vector& vector, const float& angle)
     {
-        double radians = angle * DEG_TO_RAD;
-        double sinAngle = sin(radians);
-        double cosAngle = cos(radians);
+        float radians = angle * DEG_TO_RAD;
+        float sinAngle = sin(radians);
+        float cosAngle = cos(radians);
 
         Vector newVector = {0,0};
         newVector.x = cosAngle * vector.x - sinAngle * vector.y;
@@ -48,7 +48,7 @@ public:
         return newVector;
     }
 
-    static Vector rotateVector(const Vector& vector, const double& sinAngle, const double& cosAngle)
+    static Vector rotateVector(const Vector& vector, const float& sinAngle, const float& cosAngle)
     {
         Vector newVector = {0,0};
         newVector.x = cosAngle * vector.x - sinAngle * vector.y;
@@ -58,7 +58,7 @@ public:
 
     void normalize()
     {
-        double magnitude = sqrt(x * x + y * y);
+        float magnitude = sqrt(x * x + y * y);
 
         if (magnitude <= 0.0)
             return;
@@ -67,7 +67,7 @@ public:
         y /= magnitude;
     }
 
-    void scale(const double& scale)
+    void scale(const float& scale)
     {
         x *= scale;
         y *= scale;
@@ -83,7 +83,7 @@ public:
         return {x - a.x, y - a.y};
     }
 
-    Vector operator*(const double scale) const
+    Vector operator*(const float scale) const
     {
         return {x * scale, y * scale};
     }
